requirements.txt
celery
docker-py
redis
